@@ -12,72 +12,72 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-mdc/web": "5.1.1",
-    "@angular/animations": "9.0.5",
-    "@angular/cdk": "9.1.1",
-    "@angular/common": "9.0.5",
-    "@angular/compiler": "9.0.5",
-    "@angular/core": "9.0.5",
-    "@angular/elements": "9.0.5",
-    "@angular/flex-layout": "9.0.0-beta.29",
-    "@angular/forms": "9.0.5",
-    "@angular/localize": "9.0.5",
-    "@angular/material": "9.1.1",
-    "@angular/material-moment-adapter": "9.1.1",
-    "@angular/platform-browser": "9.0.5",
-    "@angular/platform-browser-dynamic": "9.0.5",
-    "@angular/router": "9.0.5",
-    "@fortawesome/angular-fontawesome": "0.6.0",
-    "@fortawesome/fontawesome-svg-core": "1.2.27",
-    "@fortawesome/free-solid-svg-icons": "5.12.1",
-    "@ngx-i18nsupport/tooling": "8.0.3",
-    "angular2-template-loader": "0.6.2",
-    "awesome-typescript-loader": "5.2.1",
-    "core-js": "3.6.4",
-    "css-loader": "3.4.2",
-    "document-register-element": "1.14.3",
-    "html-loader": "0.5.5",
-    "mini-css-extract-plugin": "0.9.0",
-    "moment": "2.24.0",
-    "ngx-mat-select-search": "2.1.2",
-    "ngx-order-pipe": "2.0.4",
-    "npm-check-updates": "4.0.4",
-    "postcss-loader": "3.0.0",
-    "require": "2.4.20",
-    "rxjs": "6.5.4",
-    "style-loader": "1.1.3",
-    "typescript": "3.7.5",
-    "uuid": "7.0.2",
-    "zone.js": "0.10.2"
+    "@angular-mdc/web": "^5.1.1",
+    "@angular/animations": "^9.1.13",
+    "@angular/cdk": "^9.2.4",
+    "@angular/common": "^9.1.13",
+    "@angular/compiler": "^9.1.13",
+    "@angular/core": "^9.1.13",
+    "@angular/elements": "^9.1.13",
+    "@angular/flex-layout": "^9.0.0-beta.31",
+    "@angular/forms": "^9.1.13",
+    "@angular/localize": "^9.1.13",
+    "@angular/material": "^9.2.4",
+    "@angular/material-moment-adapter": "^9.2.4",
+    "@angular/platform-browser": "^9.1.13",
+    "@angular/platform-browser-dynamic": "^9.1.13",
+    "@angular/router": "^9.1.13",
+    "@fortawesome/angular-fontawesome": "^0.6.1",
+    "@fortawesome/fontawesome-svg-core": "^1.2.35",
+    "@fortawesome/free-solid-svg-icons": "^5.15.3",
+    "@ngx-i18nsupport/tooling": "^8.0.3",
+    "angular2-template-loader": "^0.6.2",
+    "awesome-typescript-loader": "^5.2.1",
+    "core-js": "^3.10.1",
+    "css-loader": "^3.6.0",
+    "document-register-element": "^1.14.10",
+    "html-loader": "^0.5.5",
+    "mini-css-extract-plugin": "^0.9.0",
+    "moment": "^2.29.1",
+    "ngx-mat-select-search": "^2.2.0",
+    "ngx-order-pipe": "^2.1.1",
+    "npm-check-updates": "^4.1.2",
+    "postcss-loader": "^3.0.0",
+    "require": "^2.4.20",
+    "rxjs": "^6.6.7",
+    "style-loader": "^1.3.0",
+    "typescript": "<3.9.0",
+    "uuid": "^7.0.3",
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.900.5",
-    "@angular/cli": "9.0.5",
-    "@angular/compiler-cli": "9.0.5",
-    "@angular/language-service": "9.0.5",
-    "@ngx-i18nsupport/ngx-i18nsupport": "1.1.6",
-    "@types/jasmine": "3.5.8",
-    "@types/jasminewd2": "2.0.8",
-    "@types/lodash": "4.14.149",
-    "@types/node": "13.9.0",
-    "extract-loader": "5.0.1",
-    "gulp": "4.0.2",
-    "gulp-clean": "0.4.0",
-    "gulp-concat": "2.6.1",
-    "gulp-copy": "4.0.1",
-    "gulp-sass": "4.0.2",
-    "gulp-sass-glob": "1.1.0",
-    "helpers": "0.0.6",
-    "jasmine-core": "3.5.0",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "4.4.1",
-    "karma-chrome-launcher": "3.1.0",
-    "karma-coverage-istanbul-reporter": "2.1.1",
-    "karma-jasmine": "3.1.1",
-    "karma-jasmine-html-reporter": "1.5.2",
-    "protractor": "5.4.3",
-    "resolve-url-loader": "3.1.1",
-    "ts-node": "8.6.2",
-    "webpack": "4.42.0"
+    "@angular-devkit/build-angular": "^0.900.7",
+    "@angular/cli": "^9.1.15",
+    "@angular/compiler-cli": "^9.1.13",
+    "@angular/language-service": "^9.1.13",
+    "@ngx-i18nsupport/ngx-i18nsupport": "^1.1.6",
+    "@types/jasmine": "^3.6.9",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/lodash": "^4.14.168",
+    "@types/node": "^13.13.48",
+    "extract-loader": "^5.1.0",
+    "gulp": "^4.0.2",
+    "gulp-clean": "^0.4.0",
+    "gulp-concat": "^2.6.1",
+    "gulp-copy": "^4.0.1",
+    "gulp-sass": "^4.1.0",
+    "gulp-sass-glob": "^1.1.0",
+    "helpers": "^0.0.6",
+    "jasmine-core": "^3.7.1",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-coverage-istanbul-reporter": "^2.1.1",
+    "karma-jasmine": "^3.3.1",
+    "karma-jasmine-html-reporter": "^1.5.4",
+    "protractor": "^5.4.4",
+    "resolve-url-loader": "^3.1.2",
+    "ts-node": "^8.10.2",
+    "webpack": "^4.46.0"
   }
 }
