@@ -1,24 +1,24 @@
-import { 
-    Component, 
+import {
+    Component,
     Inject
 } from '@angular/core';
-import { 
-    MatDialogRef, 
-    MAT_DIALOG_DATA 
+import {
+    MatDialogRef,
+    MAT_DIALOG_DATA
 } from '@angular/material/dialog';
 
 @Component({
     templateUrl: './templates/errordialog.html'
 })
-export class ErrorDialog
-{
+export class ErrorDialog {
     public text: string;
-    
+
     constructor(
-        public dialogRef : MatDialogRef<ErrorDialog>,
+        public dialogRef: MatDialogRef < ErrorDialog > ,
         @Inject(MAT_DIALOG_DATA) public message: string) {};
-    
+
     cancel() {
         this.dialogRef.close();
     };
 }
+
