@@ -1,30 +1,31 @@
-import { Injectable } from '@angular/core';
-import { 
-    Observable, 
-    BehaviorSubject 
+import {
+    Injectable
+} from '@angular/core';
+import {
+    Observable,
+    BehaviorSubject
 } from 'rxjs';
 
 @Injectable({
-    providedIn : 'root'
+    providedIn: 'root'
 })
-export class LoaderService 
-{
-    private status: BehaviorSubject<boolean>;
+export class LoaderService {
+    private status: BehaviorSubject < boolean > ;
 
     constructor() {
-        this.status = new BehaviorSubject<boolean>(false);
+        this.status = new BehaviorSubject < boolean > (false);
     }
-    
-    public getStatus(): Observable<boolean> {
+
+    public getStatus(): Observable < boolean > {
         return this.status.asObservable();
     }
-    
+
     public visible(): void {
         setTimeout(() => {
             this.status.next(true);
         });
     }
-    
+
     public invisible(): void {
         setTimeout(() => {
             this.status.next(false);
