@@ -1,19 +1,21 @@
 import moment from "moment";
 
-export abstract class BaseModel
-{
-    constructor(model?: {}) {
-        this.fillModel(model ? model : {});
-    } 
-    
-    abstract fillModel(_model: {}): {};
-    abstract getArray(): {};
-    
-    setFieldValue(value: any, entry: any) {
+export abstract class BaseModel {
+    constructor(model ? : {}) {
+        this.fillModelFromData(model ? model : {});
+    }
+
+    abstract fillModelFromData(_model: {}): {};
+
+    protected setFieldValue(value: any, entry: any) {
         return (undefined === value) ? entry : value;
     }
-    
-    setDateValue(value: any, entry: any) {
+
+    protected setDateValue(value: any, entry: any) {
         return (undefined === value) ? entry : moment(value);
+    }
+    
+    public getModelData(): object {
+        return Object.assign({}, this);
     }
 }
