@@ -1,4 +1,4 @@
-import { 
+import {
     Component,
     Inject,
     AfterViewInit,
@@ -9,8 +9,12 @@ import {
     MatDialogRef,
     MAT_DIALOG_DATA
 } from '@angular/material/dialog';
-import { MatTableDataSource } from '@angular/material/table';
-import { MatSort } from '@angular/material/sort';
+import {
+    MatTableDataSource
+} from '@angular/material/table';
+import {
+    MatSort
+} from '@angular/material/sort';
 import {
     FormBuilder,
     FormGroup,
@@ -20,52 +24,65 @@ import {
     faEdit,
     faTrash
 } from '@fortawesome/free-solid-svg-icons';
-import { BehaviorSubject } from 'rxjs';
-import { DialogService } from '@app/common/dialogs/dialog.service';
-import { APP_SELECTORS, APP_TRANSFER_TYPE } from '@app/common/appsettings/appsettings';
-import { SortingService } from '@app/common/sorting/sorting.service';
-import { TransferService } from './transfer.service';
-import { Transfer } from './transfer.model';
+import {
+    BehaviorSubject
+} from 'rxjs';
+import {
+    DialogService
+} from '@app/common/dialogs/dialog.service';
+import {
+    APP_SELECTORS,
+    APP_TRANSFER_TYPE
+} from '@app/common/appsettings/appsettings';
+import {
+    SortingService
+} from '@app/common/sorting/sorting.service';
+import {
+    TransferService
+} from './transfer.service';
+import {
+    Transfer
+} from './transfer.model';
 
 @Component({
     selector: APP_SELECTORS.TRANSFER,
     templateUrl: './templates/transfer.list.html'
 })
-export class TransferComponent implements AfterViewInit
-{
+export class TransferComponent implements AfterViewInit {
     public updateIcon = faEdit;
     public deleteIcon = faTrash;
-    private isSubmitted: BehaviorSubject<any>;
-    private dialogRef: MatDialogRef<TransferDialog>;
+    private isSubmitted: BehaviorSubject < any > ;
+    private dialogRef: MatDialogRef < TransferDialog > ;
     public displayedColumns: string[] = [
         'hardware_name',
         'type',
         'date',
         'remarks',
     ];
-    public transferList = new MatTableDataSource<Transfer>();
+    public transferList = new MatTableDataSource < Transfer > ();
 
-    @ViewChild(MatSort, {static: false}) sort: MatSort;
+    @ViewChild(MatSort, {
+        static: false
+    }) sort: MatSort;
 
     constructor(
         private dialogService: DialogService,
         private transferService: TransferService,
         private sortingService: SortingService,
         private updateDialog: MatDialog
-    ) { }
-    
+    ) {}
+
     ngAfterViewInit(): void {
         this.transferService
-            .get()
+            .list()
             .subscribe((result: Transfer[]) => {
                 this.transferList.data = result;
                 this.transferList.sort = this.sort;
                 this.transferList.sortData = this.sortingService.sort;
             });
     }
-    
-    delete(transfer: Transfer) 
-    {
+
+    delete(transfer: Transfer) {
         this.dialogService
             .confirm('Potwierdź usunięcie')
             .subscribe((confirm: boolean) => {
@@ -73,16 +90,18 @@ export class TransferComponent implements AfterViewInit
                     this.transferService
                         .destroy(transfer)
                         .subscribe(() => {
-                            let index = this.transferList.data.findIndex((item: Transfer) => {return item.id === transfer.id});
+                            let index = this.transferList.data.findIndex((item: Transfer) => {
+                                return item.id === transfer.id
+                            });
                             this.transferList.data.splice(index, 1);
                             this.transferList.data = [...this.transferList.data];
-                    });
-            }
-        });     
+                        });
+                }
+            });
     }
 
     edit(transfer: Transfer) {
-        this.isSubmitted = new BehaviorSubject<any>({});
+        this.isSubmitted = new BehaviorSubject < any > ({});
         this.isSubmitted
             .asObservable()
             .subscribe((value: any) => {
@@ -92,20 +111,22 @@ export class TransferComponent implements AfterViewInit
             });
         this.openDialog(transfer);
     }
-    
+
     update(transfer: Transfer) {
         this.transferService
             .update(transfer)
             .subscribe(() => {
                 this.closeDialog();
-                let index = this.transferList.data.findIndex((item: Transfer) => { return item.id === transfer.id });
+                let index = this.transferList.data.findIndex((item: Transfer) => {
+                    return item.id === transfer.id
+                });
                 this.transferList.data[index] = transfer;
                 this.transferList.data = [...this.transferList.data];
             });
     }
 
     create() {
-        this.isSubmitted = new BehaviorSubject<any>({});
+        this.isSubmitted = new BehaviorSubject < any > ({});
         this.isSubmitted
             .asObservable()
             .subscribe((value: any) => {
@@ -115,7 +136,7 @@ export class TransferComponent implements AfterViewInit
             });
         this.openDialog(new Transfer());
     }
-    
+
     store(transfer: Transfer) {
         this.transferService
             .store(transfer)
@@ -124,9 +145,9 @@ export class TransferComponent implements AfterViewInit
                 this.transferList.data = [...this.transferList.data, new Transfer(value)];
             });
     }
-    
+
     openDialog(transfer: Transfer) {
-       this.dialogRef = this.updateDialog.open(TransferDialog, {
+        this.dialogRef = this.updateDialog.open(TransferDialog, {
             data: {
                 model: transfer,
                 isSubmitted: this.isSubmitted
@@ -135,16 +156,16 @@ export class TransferComponent implements AfterViewInit
             disableClose: true
         });
     }
-    
+
     closeDialog() {
         this.dialogRef.close();
     }
-    
+
     applyFilter(event: Event) {
         const filterValue = (event.target as HTMLInputElement).value;
         this.transferList.filter = filterValue.trim().toLowerCase();
     }
-    
+
     getTransferTypeName(transfer: Transfer) {
         return (transfer.type === APP_TRANSFER_TYPE.LEASE) ? 'Transfer to user' : 'Return to owner';
     }
@@ -153,59 +174,54 @@ export class TransferComponent implements AfterViewInit
 @Component({
     templateUrl: './templates/transfer.update.html'
 })
-export class TransferDialog 
-{
+export class TransferDialog {
     public transferForm: FormGroup;
-    private isSubmitted: BehaviorSubject<any>;
+    private isSubmitted: BehaviorSubject < any > ;
 
     constructor(
         private formBuilder: FormBuilder,
-        private dialogRef : MatDialogRef<TransferDialog>,
+        private dialogRef: MatDialogRef < TransferDialog > ,
         @Inject(MAT_DIALOG_DATA) public data: any
     ) {
         this.transferForm = this.formBuilder.group({
             'id': [
                 this.data.model.id,
                 Validators.required
-                ],
+            ],
             'user_id': [
                 this.data.model.user_id,
                 Validators.required
-                ],
+            ],
             'hardware_id': [
                 this.data.model.hardware_id,
                 Validators.required
-                ],
+            ],
             'type': [
                 this.data.model.type,
                 Validators.required
-                ],
+            ],
             'date': [
                 this.data.model.date,
                 Validators.required
-                ],
+            ],
             'remarks': [
                 this.data.model.remarks,
                 Validators.required
-                ],
+            ],
         });
         this.isSubmitted = this.data.isSubmitted;
     };
-    
-    cancel() 
-    {
+
+    cancel() {
         this.dialogRef.close();
     };
-    
-    noValidate(field: string): boolean 
-    {
+
+    noValidate(field: string): boolean {
         return !this.transferForm.get(field).valid && this.transferForm.get(field).touched;
     }
-    
-    onSubmit(value: any) 
-    {
+
+    onSubmit(value: any) {
         let transfer = new Transfer(value);
         this.isSubmitted.next(transfer);
     }
 }
-
