@@ -1,84 +1,40 @@
-import { HttpClient } from '@angular/common/http';
-import { Injectable } from '@angular/core';
-import { finalize, map } from 'rxjs/internal/operators';
-import { LoaderService } from '@app/common/loader/loader.service';
-import { APP_ROUTES } from '@app/common/appsettings/appsettings';
-import { Transfer } from './transfer.model';
+import {
+    Injectable
+} from '@angular/core';
+import {
+    map
+} from 'rxjs/internal/operators';
+import {
+    APP_ROUTES
+} from '@app/common/appsettings/appsettings';
+import {
+    Transfer
+} from './transfer.model';
+import { 
+    GeneralService 
+} from '@app/common/services/general.service';
 
 @Injectable({
-  providedIn: 'root',
+    providedIn: 'root',
 })
-export class TransferService
-{
-    constructor(
-        private http: HttpClient, 
-        private loaderService: LoaderService
-    ) {
+export class TransferService extends GeneralService {
+    list(): any {
+        return this.get(APP_ROUTES.TRANSFER)
+            .pipe(map((response: Response) => {
+                return response['result'].map((data: any) => (new Transfer(data)));
+            }));
     }
-        
-    get(): any 
-    {
-        this.loaderService
-            .visible();
-        return this.http
-            .get<any>(APP_ROUTES.TRANSFER)
-            .pipe(map(response => {
-                    return response.result.map((data:any) => (new Transfer(data)));
-                })
-            )
-            .pipe(
-                finalize(() => {
-                    this.loaderService
-                        .invisible();
-                })
-            );
-            
-        
+
+    store(transfer: Transfer): any {
+        return this.post(APP_ROUTES.TRANSFER, transfer.getModelData())
+           .pipe(map((response: Response) => { return new Transfer(response['result']); }));
     }
-    
-    store(transfer: Transfer): any 
-    {
-        this.loaderService
-            .visible();
-        return this.http
-            .post<any>(APP_ROUTES.TRANSFER, transfer.getArray())
-            .pipe(map(response => {
-                    return new Transfer(response.result);
-                })
-            )
-            .pipe(
-                finalize(() => {
-                    this.loaderService
-                        .invisible();
-                })
-            );
-    }     
-    
-    destroy(transfer: Transfer): any 
-    {
-        this.loaderService
-            .visible();
-        return this.http
-            .delete(APP_ROUTES.TRANSFER + '/' + transfer.id)
-            .pipe(
-                finalize(() => {
-                    this.loaderService
-                        .invisible();
-                })
-            );
-    }   
-    
-    update(transfer: Transfer): any 
-    {
-        this.loaderService
-            .visible();
-        return this.http
-            .put(`${APP_ROUTES.TRANSFER}/${transfer.id}`, transfer.getArray())
-            .pipe(
-                finalize(() => {
-                    this.loaderService
-                        .invisible();
-                })
-            );
-    }     
+
+    destroy(transfer: Transfer): any {
+        return this.delete(APP_ROUTES.TRANSFER, transfer.id);
+    }
+
+    update(transfer: Transfer): any {
+        return this.put(APP_ROUTES.TRANSFER, transfer.id, transfer.getModelData());
+    }
 }
