@@ -1,84 +1,40 @@
-import { HttpClient } from '@angular/common/http';
-import { Injectable } from '@angular/core';
-import { finalize, map } from 'rxjs/internal/operators';
-import { LoaderService } from '@app/common/loader/loader.service';
-import { APP_ROUTES } from '@app/common/appsettings/appsettings';
-import { User } from './user.model';
+import {
+    Injectable
+} from '@angular/core';
+import {
+    map
+} from 'rxjs/internal/operators';
+import {
+    APP_ROUTES
+} from '@app/common/appsettings/appsettings';
+import {
+    User
+} from './user.model';
+import { 
+    GeneralService 
+} from '@app/common/services/general.service';
 
 @Injectable({
-  providedIn: 'root',
+    providedIn: 'root',
 })
-export class UserService
-{
-    constructor(
-        private http: HttpClient, 
-        private loaderService: LoaderService
-    ) {
+export class UserService extends GeneralService {
+    list(): any {
+        return this.get(APP_ROUTES.USER)
+            .pipe(map((response: Response) => {
+                return response['result'].map((data: any) => (new User(data)));
+            }));
     }
-        
-    get(): any 
-    {
-        this.loaderService
-            .visible();
-        return this.http
-            .get<any>(APP_ROUTES.USER)
-            .pipe(map(response => {
-                    return response.result.map((data:any) => (new User(data)));
-                })
-            )
-            .pipe(
-                finalize(() => {
-                    this.loaderService
-                        .invisible();
-                })
-            );
-            
-        
+
+    destroy(user: User): any {
+        return this.delete(APP_ROUTES.USER, user.id);
     }
-    
-    destroy(user: User): any 
-    {
-        this.loaderService
-            .visible();
-        return this.http
-            .delete(APP_ROUTES.USER + '/' + user.id)
-            .pipe(
-                finalize(() => {
-                    this.loaderService
-                        .invisible();
-                })
-            );
-    }   
-    
-    update(user: User): any 
-    {
-        this.loaderService
-            .visible();
-        return this.http
-            .put(`${APP_ROUTES.USER}/${user.id}`, user.getArray())
-            .pipe(
-                finalize(() => {
-                    this.loaderService
-                        .invisible();
-                })
-            );
-    } 
-    
-    store(user: User): any 
-    {
-        this.loaderService
-            .visible();
-        return this.http
-            .post<any>(APP_ROUTES.USER, user.getArray())
-            .pipe(map(response => {
-                    return new User(response.result);
-                })
-            )
-            .pipe(
-                finalize(() => {
-                    this.loaderService
-                        .invisible();
-                })
-            );
-    }     
+
+    update(user: User): any {
+        return this.put(APP_ROUTES.USER, user.id, user.getModelData());
+    }
+
+    store(user: User): any {
+        return this.post(APP_ROUTES.USER, user.getModelData())
+           .pipe(map((response: Response) => { return new User(response['result']); }));
+    }
 }
