@@ -12,5 +12,5 @@
 <body>
     <app-loader></app-loader>
     <app-root></app-root>
-<script src="js/runtime-es2015.04b153c2e37b5a90c9eb.js" type="module"></script><script src="js/runtime-es5.04b153c2e37b5a90c9eb.js" nomodule defer></script><script src="js/polyfills-es5.917c9895320d5a53bc5d.js" nomodule defer></script><script src="js/polyfills-es2015.517287b5c9977318d26e.js" type="module"></script><script src="js/scripts.9ddf727ecc94cd123172.js" defer></script><script src="js/main-es2015.92078d17b7226f0fe02c.js" type="module"></script><script src="js/main-es5.92078d17b7226f0fe02c.js" nomodule defer></script></body>
+<script src="js/runtime-es2015.js" type="module"></script><script src="js/runtime-es5.js" nomodule defer></script><script src="js/polyfills-es5.js" nomodule defer></script><script src="js/polyfills-es2015.js" type="module"></script><script src="js/scripts.js" defer></script><script src="js/vendor-es2015.js" type="module"></script><script src="js/vendor-es5.js" nomodule defer></script><script src="js/main-es2015.js" type="module"></script><script src="js/main-es5.js" nomodule defer></script></body>
 </html>
